requirements.txt
huggingface-hub<0.8
python-dotenv
streamlit==1.10.0
datasets<2.3
evaluate<0.2
jsonlines
typer[rich]
# Dataset specific deps
py7zr<0.19
openpyxl<3.1
# Dirty bug from Google
protobuf<=3.20.1